CORE-9171: moved the submission template tests into the jobs package
